error-lens-mas/error_lens_agent/requirements.txt
google-adk[a2a,extensions]>=1.28.1
python-dotenv>=1.2.2
